requirements.txt
dash==2.5.0
mypy~=0.761
pandas==1.4.2
plotly==5.8.1
PyYAML~=5.3
# slackclient~=2.5.0
slack_sdk==3.17.0